fix(auth): redirect already-authenticated visitors away from login page
LoginPage was a client component that always rendered the login form,
so users with a valid session cookie who navigated to "/" saw the form
and assumed their session had expired. Other routes under /(app)/* were
unaffected because middleware short-circuited authenticated requests.

Convert page.tsx into a server component that calls getServerSession
and redirects to /dashboard when a session exists. The form itself is
extracted unchanged into src/app/_components/login-form.tsx and only
rendered for unauthenticated visitors. Session lifetime, cookie config
and middleware semantics are unchanged.
